roachtest: move some costFuzz functions in preparation for refactoring
Release note: None
